moving og tool to it own folder outside tooling folder
