+++ original_tool/load_go_table.py
